backend/quiz-c-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
groq==0.36.0
python-decouple==3.8